refactor: 패키지 구조 변경, Service 인터페이스 제거
