+++ src/main/java/numble/daangnservice/domain/product/service/ProductService.java
@@ -6,9 +6,9 @@ import numble.daangnservice.domain.product.ProductCategory;
 import numble.daangnservice.domain.product.ProductEntity;
 import numble.daangnservice.domain.product.ProductImageEntity;
 import numble.daangnservice.domain.product.ProductStatus;
-import numble.daangnservice.domain.repository.ProductImageRepository;
-import numble.daangnservice.domain.repository.ProductRepository;
-import numble.daangnservice.domain.repository.UserRepository;
+import numble.daangnservice.repository.ProductImageRepository;
+import numble.daangnservice.repository.ProductRepository;
+import numble.daangnservice.repository.UserRepository;
 import numble.daangnservice.domain.user.UserEntity;
 import numble.daangnservice.domain.utils.UploadService;
 import numble.daangnservice.dto.ProductDto;
@@ -17,7 +17,6 @@ import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -56,7 +55,6 @@ public class ProductService {
                         .build()
         );
 
-
         for (String productImageUrl : productImageUrlsList) {
             productImageRepository.save(
                     ProductImageEntity.builder()
@@ -65,7 +63,6 @@ public class ProductService {
                             .build()
             );
         }
-
     }
 
     public ProductEntity findProduct(Long productId){
