+++ src/main/java/numble/daangnservice/dto/ProductDto.java
@@ -14,7 +14,6 @@ public class ProductDto {
     @Setter
     @NoArgsConstructor
     @AllArgsConstructor
-    @Builder
     public static class Register{
 
         @NotNull
