+++ src/main/java/numble/daangnservice/infrastructure/exception/BusinessException.java
@@ -1,4 +1,4 @@
-package numble.daangnservice.exception;
+package numble.daangnservice.infrastructure.exception;
 
 public class BusinessException extends RuntimeException{
