+++ src/main/java/numble/daangnservice/infrastructure/exception/UnknownUserException.java
@@ -1,4 +1,4 @@
-package numble.daangnservice.exception;
+package numble.daangnservice.infrastructure.exception;
 
 public class UnknownUserException extends RuntimeException {
