+++ src/main/java/numble/daangnservice/repository/ProductRepository.java
@@ -1,14 +1,12 @@
-package numble.daangnservice.domain.repository;
+package numble.daangnservice.repository;
 
 import numble.daangnservice.domain.product.ProductEntity;
 import numble.daangnservice.domain.user.UserEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
 
 import java.util.List;
 import java.util.Optional;
 
-@Repository
 public interface ProductRepository extends JpaRepository<ProductEntity, Long> {
     List<ProductEntity> findByUserEntity(UserEntity userEntity);
