+++ src/main/java/numble/daangnservice/repository/UserRepository.java
@@ -1,12 +1,10 @@
-package numble.daangnservice.domain.repository;
+package numble.daangnservice.repository;
 
 import numble.daangnservice.domain.user.UserEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
-@Repository
 public interface UserRepository extends JpaRepository<UserEntity, Long> {
 
     Optional<UserEntity> findByEmail(String email);
@@ -15,5 +13,4 @@ public interface UserRepository extends JpaRepository<UserEntity, Long> {
 
     Optional<UserEntity> findById(Long id);
 
-//    UserEntity getById(Long id);
 }
